llm_spec/requirements.txt
fastapi[standard]
uvicorn[standard]
python-multipart
pydantic
openai
python-dotenv
pandas
requests
tiktoken
numpy
groq
transformers
trl
accelerate
datasets
peft
bitsandbytes
huggingface-hub